Move test utils out vacuum
